refactor(frontend): split deployments stepper modal
